pass Order constructor test, due to changing TYPE

+++ Sales/Models/Order.cs
@@ -7,11 +7,11 @@ namespace Sales.Models
   {
     public string ProductName { get; set; }
     public string ProductDescription { get; set; }
-    public int ProductPrice { get; set; }
+    public double ProductPrice { get; set; }
     // public int OrderDate {did we do anything with dates??? look into this later.}
     private static List<Order> _products = new List<Order> {} ;
 
-    public Order(string productName, string productDescription, int productPrice)
+    public Order(string productName, string productDescription, double productPrice)
     {
       ProductName = productName;
       ProductDescription = productDescription;
